+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file. This projec
 - Remove the Site Health dashboard widget if on maintenance plan.
 - Remove the help tab.
 - Remove import/export pages unless `BM_WP_ENABLE_IMPORT_EXPORT` is set to `true`.
+- Custom branded admin theme.
 
 ### Changed